CP-12617: Port CP-12191 to trunk
Add a XenCenter option to enable the old behaviour for the default search for servers
