+++ XenAdmin/Dialogs/OptionsDialog.resx
@@ -183,6 +183,36 @@
   <data name="&gt;&gt;saveAndRestoreOptionsPage1.ZOrder" xml:space="preserve">
     <value>1</value>
   </data>
+  <data name="searchOptionsPage1.AutoSize" type="System.Boolean, mscorlib">
+    <value>True</value>
+  </data>
+  <data name="searchOptionsPage1.Dock" type="System.Windows.Forms.DockStyle, System.Windows.Forms">
+    <value>Fill</value>
+  </data>
+  <data name="searchOptionsPage1.Font" type="System.Drawing.Font, System.Drawing">
+    <value>Segoe UI, 9pt</value>
+  </data>
+  <data name="searchOptionsPage1.Location" type="System.Drawing.Point, System.Drawing">
+    <value>10, 49</value>
+  </data>
+  <data name="searchOptionsPage1.Size" type="System.Drawing.Size, System.Drawing">
+    <value>534, 451</value>
+  </data>
+  <data name="searchOptionsPage1.TabIndex" type="System.Int32, mscorlib">
+    <value>6</value>
+  </data>
+  <data name="&gt;&gt;searchOptionsPage1.Name" xml:space="preserve">
+    <value>searchOptionsPage1</value>
+  </data>
+  <data name="&gt;&gt;searchOptionsPage1.Type" xml:space="preserve">
+    <value>XenAdmin.Dialogs.OptionsPages.searchOptionsPage, XenCenterMain, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null</value>
+  </data>
+  <data name="&gt;&gt;searchOptionsPage1.Parent" xml:space="preserve">
+    <value>ContentPanel</value>
+  </data>
+  <data name="&gt;&gt;searchOptionsPage1.ZOrder" xml:space="preserve">
+    <value>1</value>
+  </data>
   <data name="securityOptionsPage1.Dock" type="System.Windows.Forms.DockStyle, System.Windows.Forms">
     <value>Fill</value>
   </data>
