+++ XenAdmin/app.config
@@ -17,7 +17,8 @@
             </setting>
             <setting name="ServerList" serializeAs="Xml">
                 <value>
-                    <ArrayOfString xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema"/>
+                    <ArrayOfString xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                        xmlns:xsd="http://www.w3.org/2001/XMLSchema" />
                 </value>
             </setting>
             <setting name="LocalSRsVisible" serializeAs="String">
@@ -57,7 +58,7 @@
                 <value>0</value>
             </setting>
             <setting name="ProxyAddress" serializeAs="String">
-                <value/>
+                <value />
             </setting>
             <setting name="BypassProxyForLocal" serializeAs="String">
                 <value>False</value>
@@ -102,7 +103,7 @@
                 <value>False</value>
             </setting>
             <setting name="LatestXenCenterSeen" serializeAs="String">
-                <value/>
+                <value />
             </setting>
             <setting name="SeenAllowUpdatesDialog" serializeAs="String">
                 <value>False</value>
@@ -118,16 +119,18 @@
             </setting>
             <setting name="DisabledPlugins" serializeAs="Xml">
                 <value>
-                    <ArrayOfString xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema"/>
+                    <ArrayOfString xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                        xmlns:xsd="http://www.w3.org/2001/XMLSchema" />
                 </value>
             </setting>
             <setting name="IgnoreFirstRunWizards" serializeAs="Xml">
                 <value>
-                    <ArrayOfString xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema"/>
+                    <ArrayOfString xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                        xmlns:xsd="http://www.w3.org/2001/XMLSchema" />
                 </value>
             </setting>
             <setting name="ServerStatusPath" serializeAs="String">
-                <value/>
+                <value />
             </setting>
             <setting name="RollingUpgradeWizardShowFirstPage" serializeAs="String">
                 <value>False</value>
@@ -136,7 +139,7 @@
                 <value>True</value>
             </setting>
             <setting name="ApplicationVersion" serializeAs="String">
-                <value/>
+                <value />
             </setting>
             <setting name="UncaptureShortcutKey" serializeAs="String">
                 <value>0</value>
@@ -149,6 +152,9 @@
             </setting>
             <setting name="ShowHealthCheckEnrollmentReminder" serializeAs="String">
                 <value>True</value>
+			</setting>
+            <setting name="ShowJustHostInSearch" serializeAs="String">
+                <value>False</value>
             </setting>
         </XenAdmin.Properties.Settings>
     </userSettings>
